# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PIX Payment Integration API",
    "description": "Complete PIX (Brazilian instant payment) integration for Love to Fly Portal. Supports hangar bookings and membership upgrades with webhook and reconciliation mechanisms.",
    "version": "1.0.0",
    "contact": {
      "name": "Love to Fly Team",
      "email": "support@lovetofly.com"
    }
  },
  "servers": [
    {
      "url": "https://lovetofly.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from user authentication"
      },
      "CronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-secret",
        "description": "Secret for cron job reconciliation"
      },
      "WebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "asaas-access-token",
        "description": "Webhook token configured in Asaas (asaas-access-token header)"
      }
    },
    "schemas": {
...
      "PaymentStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "order_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "completed", "expired", "failed"]
          },
          "amount": {
            "type": "object",
            "properties": {
              "cents": {
                "type": "integer"
              },
              "formatted": {
                "type": "string",
                "example": "R$ 500,00"
              }
            }
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "payment_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "ReconcileTransaction": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string"
          },
          "txid": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["paid", "confirmed", "settled", "pending", "expired"]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/payments/pix/status/{orderId}": {
      "get": {
        "tags": ["PIX Payment Status"],
        "summary": "Get payment status",
        "description": "Check current status of a PIX payment. Used for polling after QR code is scanned.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order reference ID (e.g., 'booking-123')"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatus"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/payments/pix/reconcile": {
      "post": {
        "tags": ["PIX Reconciliation"],
        "summary": "Bulk reconciliation of transactions",
        "description": "Cron-friendly endpoint for reconciling transactions from Asaas API. Can be called periodically to sync payment status.",
        "security": [
          {
            "CronSecret": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transactions": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ReconcileTransaction"
                    },
                    "description": "List of transactions from Asaas API"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reconciliation completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "processed": {
                      "type": "integer"
                    },
                    "skipped": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid cron secret"
          }
        }
      }
    },
...
    "/api/user/membership/confirm-pix-payment": {
      "post": {
        "tags": ["Membership Upgrades"],
        "summary": "Confirm membership upgrade via PIX",
        "description": "User-facing endpoint to confirm their own membership upgrade after PIX payment. Idempotent - safe to call multiple times.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId"],
                "properties": {
                  "orderId": {
                    "type": "string",
                    "example": "membership-pro",
                    "description": "Order ID for the membership upgrade"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Membership confirmation processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "membership": {
                      "type": "object",
                      "properties": {
                        "plan": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "activated_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid order or payment not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
...
}